--- app/src/main/java/com/interview/project/ui/utils/MyAppGlideModule.kt
@@ -1,12 +1,0 @@
-package com.interview.project.ui.utils
-
-import com.bumptech.glide.annotation.GlideModule
-import com.bumptech.glide.module.AppGlideModule
-
-/**
- * Created by akshay on 26,October,2020
- * akshay2211@github.io
- */
-
-@GlideModule
-class MyAppGlideModule : AppGlideModule()
