--- app/src/main/java/com/interview/project/ui/utils/SquareImageView.kt
@@ -1,56 +1,0 @@
-package com.interview.project.ui.utils
-
-import android.content.Context
-import android.graphics.Canvas
-import android.graphics.Path
-import android.graphics.RectF
-import android.util.AttributeSet
-import androidx.appcompat.widget.AppCompatImageView
-
-class SquareImageView @JvmOverloads constructor(
-    context: Context?,
-    attrs: AttributeSet? = null,
-    defStyleAttr: Int = 0
-) : AppCompatImageView(
-    context!!, attrs, defStyleAttr
-) {
-
-    private val radius = 18.0f
-    private var path = Path()
-    private var rect: RectF? = null
-
-    override fun onMeasure(widthMeasureSpec: Int, heightMeasureSpec: Int) {
-        val widthSize = MeasureSpec.getSize(widthMeasureSpec)
-        val heightSize = MeasureSpec.getSize(heightMeasureSpec)
-        if (widthSize == 0 && heightSize == 0) {
-            // If there are no constraints on size, let FrameLayout measure
-            super.onMeasure(widthMeasureSpec, heightMeasureSpec)
-
-            // Now use the smallest of the measured dimensions for both dimensions
-            val minSize = Math.min(measuredWidth, measuredHeight)
-            setMeasuredDimension(minSize, minSize)
-            return
-        }
-        val size: Int
-        size = if (widthSize == 0 || heightSize == 0) {
-            // If one of the dimensions has no restriction on size, set both dimensions to be the
-            // on that does
-            widthSize.coerceAtLeast(heightSize)
-        } else {
-            // Both dimensions have restrictions on size, set both dimensions to be the
-            // smallest of the two
-            widthSize.coerceAtMost(heightSize)
-        }
-        val newMeasureSpec = MeasureSpec.makeMeasureSpec(size, MeasureSpec.EXACTLY)
-        super.onMeasure(newMeasureSpec, newMeasureSpec)
-    }
-
-    override fun onDraw(canvas: Canvas) {
-        if (rect == null) {
-            rect = RectF(0f, 0f, this.width.toFloat(), this.height.toFloat())
-        }
-        path.addRoundRect(rect!!, radius, radius, Path.Direction.CW)
-        canvas.clipPath(path)
-        super.onDraw(canvas)
-    }
-}
