+++ app/src/main/java/io/ak1/imgur/model/BaseData.kt
@@ -1,0 +1,15 @@
+package io.ak1.imgur.model
+
+import com.google.gson.annotations.SerializedName
+import java.io.Serializable
+
+/**
+ * Created by akshay on 24,October,2020
+ * akshay2211@github.io
+ */
+
+data class BaseData(
+    @SerializedName("data") var data: List<Data>? = null,
+    @SerializedName("success") var success: Boolean = false,
+    @SerializedName("status") var status: Int = 0
+) : Serializable
