+++ app/src/main/java/io/ak1/imgur/ui/utils/Constants.kt
@@ -1,0 +1,14 @@
+package io.ak1.imgur.ui.utils
+
+/**
+ * Created by akshay on 27,October,2020
+ * akshay2211@github.io
+ */
+object Constants {
+    var HeightPX = 0
+    var WidthPX = 0
+    var DENSITY = 0f
+
+    //Add Imgur Access key here
+    const val IMGUR_API_KEY = "<Access key>"
+}
