+++ app/src/main/java/io/ak1/imgur/ui/utils/MyAppGlideModule.kt
@@ -1,0 +1,12 @@
+package io.ak1.imgur.ui.utils
+
+import com.bumptech.glide.annotation.GlideModule
+import com.bumptech.glide.module.AppGlideModule
+
+/**
+ * Created by akshay on 26,October,2020
+ * akshay2211@github.io
+ */
+
+@GlideModule
+class MyAppGlideModule : AppGlideModule()
