+++ app/src/main/res/layout/images_row.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<com.interview.project.ui.utils.SquareImageView xmlns:android="http://schemas.android.com/apk/res/android"
+<io.ak1.imgur.ui.utils.SquareImageView xmlns:android="http://schemas.android.com/apk/res/android"
     android:id="@+id/imageView"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
